requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
joblib
pyyaml
fire
